+++ src/components/recipes/RecipeCard.module.css
@@ -32,21 +32,6 @@
   font-weight: 300;
 }
 
-.recipeBtn {
-  font-size: 30px;
-  justify-self: flex-end;
-  margin-left: 15px;
-}
-
-.savedRecipeBtn,
-.unsavedRecipeBtn:hover {
-  color: #1cd22e;
-}
-
-.savedRecipeBtn:hover {
-  color: #8ceb95;
-}
-
 .recipeImageContainer {
   position: relative;
   width: 225px;
